refactor guide file structure
