+++ src/app/(admin)/new_dashboard/guides/hooks/useSearchGuides.ts
@@ -1,0 +1,20 @@
+import { useState } from "react";
+import { Guide } from "../types";
+
+export const useSearchGuides = (initialGuides: Guide[]) => {
+  const [searchQuery, setSearchQuery] = useState("");
+
+  const filteredGuides = initialGuides.filter((guide) =>
+    guide.title.toLowerCase().includes(searchQuery.toLowerCase())
+  );
+
+  const handleSearchChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setSearchQuery(event.target.value);
+  };
+
+  return {
+    searchQuery,
+    filteredGuides,
+    handleSearchChange,
+  };
+};
