+++ src/app/(admin)/new_dashboard/guides/types/index.ts
@@ -1,0 +1,6 @@
+export type Guide = {
+  title: string;
+  issue: string;
+  likes: number;
+  dislikes: number;
+};
